home-manager-tool: prioritize `-I` parameters
In 176e455 the order between the action of `-I` parameters getting
added to `EXTRA_NIX_PATH` and the action of a static path getting
added to `EXTRA_NIX_PATH` was reversed, also reversing the order of
`-I` parameters and the static `-I home-manager=...` leading to the
static `-I home-manager=...` to always come before any of the `-I`
parameters to later calls to nix commands.

This made it imposible to override the static home-manager path when
calling the home-manager tool with `-I home-manager=...`. This was
previously posible.