README: Remove the pills (#4181)
This removes the Nix Pills reference, because they are not a good
introduction to the Nix ecosystem, but rather a thorough explanation
of many disparate things. Reading through them might give some light
bulb moments for an intermediate reader, but that does not mean that
they're good for a beginner.

I've also removed the mention of infinite recursion without source
location. That's an old meme for a problem that has been mostly
solved. Mentioning it here has two effects:

- Propagate the outdated meme.

- Make users insensitive to bad errors. Learned helplessness. That
  kind of thing. What we really want is for them to report bad error
  messages, so that they can be fixed. And they can be fixed; just
  report them at `NixOS/nix`.