activation-script: default to skipping 'sanity' checks on darwin
this defaults to not checking that the `$HOME` env variable matches the
nix config entry for that user.

i think it's a safe assertion to make that given nix-darwin can't be
used for from-first-boot host provisioning, a lot less people manage
their user accounts with it. these "sanity" checks depend on the user in
question being a fully managed user, which is otherwise unnecessary and
caused some surprise breakages on home-manager updates.