bat: support boolean flags in config (fixes #4657)
Previously, users cannot enable boolean flags like `--show-all` in bat's
config since all options were expected to be either a string, or a list
of strings. With this commit boolean flags are simply appended to the
end of the config if they are set to `true`, and discarded otherwise.

For example, the config
```nix
{
  theme = "TwoDark";
  show-all = true;
  lessopen = false;
}
```
would produce a config file that looks like
```
--theme='TwoDark'
--show-all
```