hyprland: change plugins settings generation
The previous syntax (`plugin:name { ...settings}`) is not
working with more than one plugin. There is no documentation
for this, just the [source code](https://github.com/hyprwm/Hyprland/blob/main/src/config/ConfigManager.cpp#L1574).
With this update the plugins paths aren't generated together with the full config
(so the "plugin" field is not "important" anymore) and the plugins settings are generated
like the other fields.