firefox: add librewolf support
This adds LibreWolf support and tests, including using the correct
profile path and generating the correct search json hash.

The way Firefox (& forks) profiles work on Linux is you have ~/.mozilla
for some more "global" stuff like native messaging hosts, and
~/.<vendor>/<appName> for more "local" stuff like profiles or crash
reports. However, if vendor isn't set, ~/.appName is used instead.
Additionally, Firefox forks usually change ~/.mozilla to match
~/.<vendor>.

This is a huge mess so I don't expect it to work for every Firefox fork
out there - instead, I only provide support for the browser I personally
use, and if anyone wishes to add another one, they'd fine it relatively
easy to do now that all you need to do is make sure the values at the
top are consistent with where the browser stores its data.

Some of it could perhaps be moved to nixpkgs, but, again, considering
how messy this *can* be (I don't know how bad things are in practice),
this may or may not be a good idea.