vim,neovim: set EDITOR to full binary path when `defaultEditor` is used
Prior to this change the `EDITOR` variable was set only to the name of the
program that should be looked up in `PATH` to use as the editor, but this will
only work for user-level stuff, for stuff like `sudo -e` it won't work as the
`PATH` will not contain the location of such programs. This makes it so that
when the `defaultEditor` variable is set to `true` then `EDITOR` will point to
the full binary location of the `vim` and `nvim` packages accordingly.

This is similar to how we set `defaultEditor` in the Emacs module.

This change is also made for `vim`.