joplin-desktop: support profiles and more options
Joplin supports creating multiple profiles. This functionality was
previously missing from the joplin-desktop module and is implemented
in this PR.
In order to support profiles, the structure of the defined options had
to be split up into "global" settings that are shared between profiles
and profile-specific settings. This is a breaking change, but
necessary to avoid confusion about the difference between global and
profile-specific settings.

In addition to profiles, a few new options (such as markdown settings)
have been added.