thunderbird: remove Version = 2 from profiles
This fixes app startup for thunderbird installations that do not use
`wrapThunderbird` `nixpkgs` wrapper that enforces
`MOZ_LEGACY_PROFILES=1`. (This is, largely, Darwin installations from
`homebrew` or a similar mechanism.)

The fix similar to the other patch for Firefox:
https://github.com/nix-community/home-manager/pull/5724

This patch removes `Version=2` unconditionally for all platforms. This
is because regardless of whether the app is installed from `nixpkgs` or
externally, it should always use v1 behavior (with a separate
`installs.ini` file available for write.) This is already enforced for
`nixpkgs` packages with the wrapper anyway, so removing this line
shouldn't affect anyone.

Signed-off-by: Ihar Hrachyshka <ihar.hrachyshka@gmail.com>