kanshi: do not quote criteria and fix testcases
kashi doesn't require the whole criteria string to be quoted. However quoting everything
makes it impossible to add the actually required quotes around manufacturer names with spaces.

This commit also fixes the test cases that use manufacturer and model without serial as
criteria, as kanshi requires missing values like the serial to be replaced with "Unknown"