thunderbird: set MOZ_* variables for legacy profiles.ini
This fixes app startup for installations that are not using
`wrapThunderbird` (e.g. app from `homebrew`).

These variables are already set for nixpkgs thunderbird using
wrapThunderbird wrapper.

This is most helpful for Darwin users. But it will help with an
(unusual) scenario where Home Manager is used in a non-NixOS environment
with the app managed by another package manager, or installed directly
from Mozilla builds.