wpaperd: add systemd service; move to services/ from programs/
Re: #4538, wpaperd needs something to run it after being enabled. This commit
adds a systemd service to run it, and accordingly moves it to services.wpaperd.

In addition, the existing tests have been migrated to services, and an entry
in the newslist has been created alerting users to this change.