launchd: wrap agents in .app bundles
This causes agents to appear in System Settings with a fancy name (and
optionally icon), rather than the file name of the executable. This is
particularly desirous when the "executable" is a shell script that waits
for the Nix store to be available before running the true agent, because
those show up in System Settings as `sh`.