pay-respects: integrate alias configuration and fix nushell integration
This commit adds the possibility to configure the alias to use
pay-respects for all shells. The default is 'f'.
This commit also fixes the nushell integration, as pay-respects tries to
write the necessary alias command by itself into the config file. As a
solution we capture the output of the command and write that output to
the config with home-manager, thus resulting in the correct alias and
using the "correct" output of pay-respects. This might break in the
future, depending on pay-respects.