wayland: create tray.target if xsession is not enabled
As it current stands, the tray target is only created if xsession is
enabled. This causes services that depend on it to fail on wayland
sessions on configurations that do not have `services.xsession.enable =
true;` or the workaround as the following:

```nix
{config, lib, ...}: {
  systemd.user.targets.tray = lib.mkIf (!config.xsession.enable) {
    Unit = {
      Description = "Home Manager System Tray";
      Requires = [ "graphical-session-pre.target" ];
    };
  };
}
```

This PR addresses [#6329](https://github.com/nix-community/home-manager/issues/6329) and
[#2064](https://github.com/nix-community/home-manager/issues/2064),
albeit I do not think it is the greatest solution, since it allows the
targets defined by the xsession and wayland modules to be kept out of
sync.

I'd recommend defining `tray.target` independently from either one of
those modules, since virtually every graphical session that requires a
system tray would need such target.