tests: Enable firefox tests for darwin
This change also removes checks for MOZ_APP_LAUNCHER set for wrapped
package: the wrapped binary is in a different location on Darwin; but
the check seems out of place for Home Manager since it doesn't control
exact variables set by nixpkgs package.

This change also removes handling of state version < 19.09 when the
firefox package was not wrapped. It's time to clean up.

Signed-off-by: Ihar Hrachyshka <ihar.hrachyshka@gmail.com>